requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
langchain-pinecone
faiss-cpu
pydantic
pydantic-settings
python-dotenv
pandas
tiktoken
pytest
pytest-asyncio
pytest-cov
httpx
ruff
black